Simplify end to end tests execution.
Instead of diffing the output of `gorgon start` with an expected output,
it is more robust to test different output sections/hunks in the main
rspec specs. Along with robustness, it paves an easier way to integrate
end to end tests in a continuous testing environment.

The specs test only the important parts of the output hunks. It ignores
the ASCII color encodings and also backtrace in case there is any error.

The backtrace is way too much system dependent. It depends upon a
developer's machine setup, how they have installed rubies(rvm or
ruby-install). Moreover it does not provide any critical information
when testing functionality of Gorgon itself.

The end to end tests, however, fail when the entire suite is run and
those are not prepared to be run (no rabbitmq running and no gorgon
listen called in another terminal).